Implement a viable mitigation to the stuck choke announcement problem
Encountered a problem which occurs when a peer is slow to respond to a
handshake. The choke algorithm fills the peer's 'peersToUnchoke' channel
and it hangs the main thread.

We now only fill if the channel is not filled and emit a warning if the
peer is not reading the channel.